+++ experiments/RELEASES.md
@@ -9,3 +9,74 @@ Based on:
 - [python v1.0.0] experiments
 ### Releases
 - [PyPI v1.0.0] https://pypi.org/project/wingspan_experiments/1.0.0 - experiments
+
+
+## 2023-11-13 01:25:30
+### Changes
+Based on:
+- OpenAPI Doc 1.0.0 https://docs.wingspan.app/openapi/6470f38d65c260000c025474
+- Speakeasy CLI 1.119.0 (2.188.1) https://github.com/speakeasy-api/speakeasy
+### Generated
+- [python v2.0.0] experiments
+### Releases
+- [PyPI v2.0.0] https://pypi.org/project/wingspan_experiments/2.0.0 - experiments
+
+## 2023-11-20 01:26:29
+### Changes
+Based on:
+- OpenAPI Doc 1.0.0 https://docs.wingspan.app/openapi/6470f38d65c260000c025474
+- Speakeasy CLI 1.121.1 (2.194.1) https://github.com/speakeasy-api/speakeasy
+### Generated
+- [python v2.0.1] experiments
+### Releases
+- [PyPI v2.0.1] https://pypi.org/project/wingspan_experiments/2.0.1 - experiments
+
+## 2023-11-27 01:26:05
+### Changes
+Based on:
+- OpenAPI Doc 1.0.0 https://docs.wingspan.app/openapi/6470f38d65c260000c025474
+- Speakeasy CLI 1.122.0 (2.195.2) https://github.com/speakeasy-api/speakeasy
+### Generated
+- [python v2.0.2] experiments
+### Releases
+- [PyPI v2.0.2] https://pypi.org/project/wingspan_experiments/2.0.2 - experiments
+
+## 2023-12-11 01:26:50
+### Changes
+Based on:
+- OpenAPI Doc 1.0.0 https://docs.wingspan.app/openapi/6470f38d65c260000c025474
+- Speakeasy CLI 1.125.2 (2.210.6) https://github.com/speakeasy-api/speakeasy
+### Generated
+- [python v2.0.3] experiments
+### Releases
+- [PyPI v2.0.3] https://pypi.org/project/wingspan_experiments/2.0.3 - experiments
+
+## 2023-12-18 01:26:37
+### Changes
+Based on:
+- OpenAPI Doc 1.0.0 https://docs.wingspan.app/openapi/6470f38d65c260000c025474
+- Speakeasy CLI 1.126.4 (2.214.10) https://github.com/speakeasy-api/speakeasy
+### Generated
+- [python v2.1.0] experiments
+### Releases
+- [PyPI v2.1.0] https://pypi.org/project/wingspan_experiments/2.1.0 - experiments
+
+## 2024-01-08 01:26:52
+### Changes
+Based on:
+- OpenAPI Doc 1.0.0 https://docs.wingspan.app/openapi/6470f38d65c260000c025474
+- Speakeasy CLI 1.133.1 (2.228.1) https://github.com/speakeasy-api/speakeasy
+### Generated
+- [python v3.0.0] experiments
+### Releases
+- [PyPI v3.0.0] https://pypi.org/project/wingspan_experiments/3.0.0 - experiments
+
+## 2024-01-22 01:28:30
+### Changes
+Based on:
+- OpenAPI Doc 1.0.0 https://docs.wingspan.app/openapi/6470f38d65c260000c025474
+- Speakeasy CLI 1.148.0 (2.237.3) https://github.com/speakeasy-api/speakeasy
+### Generated
+- [python v3.0.1] experiments
+### Releases
+- [PyPI v3.0.1] https://pypi.org/project/wingspan_experiments/3.0.1 - experiments